feat: implement JWT authentication with access and refresh token generation, user login/logout functionality, and secure routes

+++ src/utils/cloudinary.js
@@ -1,6 +1,15 @@
+import dotenv from "dotenv"
+dotenv.config({
+  path: "./.env"
+});
 import { v2 as cloudinary } from "cloudinary";
-import { response } from "express";
-import fs from "fs"
+import fs from "fs";
+
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET
+});
 
 
 const uploadFileOnCloudinary = async (localFilePath) => {
@@ -9,19 +18,13 @@ const uploadFileOnCloudinary = async (localFilePath) => {
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto"
     });
-    console.log("File Uploaded on Cloudinary successfully : ", response.url);
+    // console.log("File Uploaded on Cloudinary successfully : ", response.url);
+    fs.unlinkSync(localFilePath)
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
+    if (fs.existsSync(localFilePath)) { fs.unlinkSync(localFilePath); }
     return null;
   }
 }
 
-
-cloudinary.config({
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET
-});
-
-export { uploadFileOnCloudinary }
+export { uploadFileOnCloudinary };
